+++ packages/graphql_codegen/lib/src/printer/keywords.dart
@@ -61,4 +61,6 @@ const staticKeywords = {
   "dynamic",
   "implements",
   "static",
+  "int",
+  "double"
 };